firmware/server/pillow/requirements.txt
paho-mqtt>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
